cloud_modules_azure/requirements.txt
azure-storage-blob==0.37.0
azure-cosmos
